Criando centidade arquivo

+++ tsemh-java/tsemh-api-rest/src/main/java/io/tsemh/tsemhapirest/repository/RegistroRepository.java
@@ -3,6 +3,6 @@ package io.tsemh.tsemhapirest.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import io.tsemh.tsemhapirest.entity.Registro;
 
-public interface RegistroRepository extends JpaRepository<Registro, Integer> {
+public interface RegistroRepository extends JpaRepository<Registro, Long> {
 
 }
